+++ blocklets/ai-studio/api/src/routes/index.ts
@@ -2,7 +2,9 @@ import { Router } from 'express';
 
 import ai from './ai';
 import { branchRoutes } from './branch';
-import datasetItems from './dataset-items';
+import datasetsItems from './dataset-items';
+import datasetItems from './dataset/item';
+import dataset from './dataset/list';
 import datasets from './datasets';
 import { globalRoutes } from './global';
 import { logRoutes } from './log';
@@ -23,8 +25,10 @@ treeRoutes(router);
 resourceRoutes(router);
 
 router.use('/ai', ai);
+router.use('/dataset', dataset);
+router.use('/dataset', datasetItems);
 router.use('/datasets', datasets);
-router.use('/datasets', datasetItems);
+router.use('/datasets', datasetsItems);
 router.use(ws);
 
 export default router;