+++ blocklets/ai-studio/api/src/routes/ws.ts
@@ -1,6 +1,7 @@
 import { user } from '@blocklet/sdk/lib/middlewares';
 import express from 'express';
 
+import { getResourceProjects } from '../libs/resource';
 import { ensurePromptsEditor, isRefReadOnly } from '../libs/security';
 import Project from '../store/models/project';
 import { getRepository } from '../store/repository';
@@ -16,7 +17,10 @@ router.use(ensurePromptsEditor, user()).ws('/ws/:projectId/:ref', async (conn, r
   const { role } = req.user!;
   const readOnly = isRefReadOnly({ ref, role });
 
-  const project = await Project.findOne({ where: { _id: projectId } });
+  const project =
+    (await Project.findOne({ where: { _id: projectId } })) ||
+    getResourceProjects('example').find((x) => x._id === projectId);
+
   if (!project) {
     conn.close(3001, `Project ${projectId} not found`);
     return;