+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.1.268 (2024-2-27)
+
+- fix: missing @blocklet/ai-builtin package
+
+## 0.1.267 (2024-2-27)
+
+- feat: add publish view
+
+## 0.1.266 (2024-2-26)
+
+- fix: support resource blocklet
+
+## 0.1.264 (2024-2-26)
+
+- feat: dataset protocol
+
 ## 0.1.263 (2024-2-23)
 
 - feat: accordion default expand